prioritize and fix virtual LED for tuna devices
In stock code, two "virtual" LED type share the same physical LED: attention
and notification.  If one is turned on and the OTHER is turned off, then the
physical LED is left in an OFF state.

This change optionally adds an additional virtual LED (charging/battery) which
is disabled at compile time by default.  As well, it then seperates all the
virtual LED's into a stack prioritized as notification, attention, and
charging (if enabled).  If a lower priority LED takes some action while a
higher priority LED is in use, the lower priority state is stored but the
physical LED is not changed.  Once the higher priority virtual LED is turned
off, the lower priority LED (if its still enabled) will take over the
physical LED.

To enable the charging virtual LED, define LIBLIGHTS_SUPPORT_CHARGING_LED as
1 (or edit the default near the top of lights.c in this commit.)

Change-Id: Icc8ed42cc797a5b1564cfcff10230a6f5e16482d
Signed-off-by: Gary Dezern <garyd9@gmail.com>